+++ benchmark/benchmark_models.py
@@ -1,4 +1,3 @@
-from benchmark.benchmark_utils import run_benchmark
 import time
 from typing import Tuple
 
@@ -218,3 +217,28 @@ def benchmark_ov_model(ov_model: ov.CompiledModel) -> OVBenchmark:
 
 def benchmark_cuda_model(cuda_model: torch.nn.Module, device: str, dtype: torch.dtype):
     run_benchmark(cuda_model, device, dtype)
+
+
+def run_benchmark(
+    model: torch.nn.Module,
+    device: str,
+    dtype: torch.dtype,
+    ort_session: ort.InferenceSession = None,
+    onnx: bool = False,
+) -> None:
+    """
+    Run and log the benchmark for the given model, device, and dtype.
+
+    :param onnx:
+    :param ort_session:
+    :param model: The model to be benchmarked.
+    :param device: The device to run the benchmark on ("cpu" or "cuda").
+    :param dtype: The data type to be used in the benchmark (typically torch.float32 or torch.float16).
+    """
+    if onnx:
+        logging.info(f"Running Benchmark for ONNX")
+        benchmark = ONNXBenchmark(ort_session, input_shape=(32, 3, 224, 224))
+    else:
+        logging.info(f"Running Benchmark for {device.upper()} and precision {dtype}")
+        benchmark = PyTorchBenchmark(model, device=device, dtype=dtype)
+    benchmark.run()
